Gem Destruction, Board Refill, Spawned Gems Lerp to postition
the player can now not move gems while other gems are being matched, destroyed, and spawned. It's just a timing system so it might be possible to spawn gems out of time with everything else and cause gems to spawn in the same position or cause null spots in the board array, not sure yet.